Conditional display of events depending on logged user

+++ src/Controller/EventController.php
@@ -10,6 +10,8 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
+use App\Security\Voter\EventVoter;
 
 #[Route('/event')]
 class EventController extends AbstractController
@@ -17,12 +19,21 @@ class EventController extends AbstractController
     #[Route(name: 'app_event_index', methods: ['GET'])]
     public function index(EventRepository $eventRepository): Response
     {
+        if ($this->isGranted('ROLE_ADMIN')) {
+            $events = $eventRepository->findAll();
+        } else {
+            $events = $eventRepository->findBy([
+                'owner' => $this->getUser(),
+            ]);
+        }
+
         return $this->render('event/index.html.twig', [
-            'events' => $eventRepository->findAll(),
+            'events' => $events,
         ]);
     }
 
     #[Route('/new', name: 'app_event_new', methods: ['GET', 'POST'])]
+    #[IsGranted('ROLE_USER')]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
         $event = new Event();
@@ -30,6 +41,12 @@ class EventController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $user = $this->getUser();
+            if (!$user) {
+                throw $this->createAccessDeniedException();
+            }
+
+            $event->setOwner($this->getUser());
             $entityManager->persist($event);
             $entityManager->flush();
 
@@ -45,6 +62,11 @@ class EventController extends AbstractController
     #[Route('/{id}', name: 'app_event_show', methods: ['GET'])]
     public function show(Event $event): Response
     {
+        $this->denyAccessUnlessGranted(
+            EventVoter::VIEW,
+            $event
+        );
+
         return $this->render('event/show.html.twig', [
             'event' => $event,
         ]);
@@ -53,6 +75,11 @@ class EventController extends AbstractController
     #[Route('/{id}/edit', name: 'app_event_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Event $event, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted(
+            EventVoter::EDIT,
+            $event
+        );
+
         $form = $this->createForm(EventType::class, $event);
         $form->handleRequest($request);
 
@@ -71,6 +98,11 @@ class EventController extends AbstractController
     #[Route('/{id}', name: 'app_event_delete', methods: ['POST'])]
     public function delete(Request $request, Event $event, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted(
+            EventVoter::DELETE,
+            $event
+        );
+
         if ($this->isCsrfTokenValid('delete'.$event->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($event);
             $entityManager->flush();
